+++ app/src/components/HelpPopup/HelpPopup.js
@@ -3,11 +3,16 @@ import { Popup } from 'semantic-ui-react'
 import HelpPopupIcon from './HelpPopupIcon'
 
 export const style = {
-  maxWidth: 417,
+  maxWidth: 465,
   padding: '2rem 1.8rem'
 }
 
-const HelpPopup = ({ children, content, trigger = <HelpPopupIcon /> }) => {
+const HelpPopup = ({
+  children,
+  content,
+  className,
+  trigger = <HelpPopupIcon className={className} />
+}) => {
   const render = content || children
   return (
     <Popup