+++ app/src/components/Trends/TrendsReChart.js
@@ -15,7 +15,7 @@ import { formatNumber } from './../../utils/formatting'
 import { mergeTimeseriesByKey } from './../../utils/utils'
 import { sourcesMeta as chartsMeta } from './trendsUtils'
 
-const ASSET_PRICE_COLOR = '#a4acb7'
+const ASSET_PRICE_COLOR = '#A4ACB7'
 
 const Loading = () => <h2 style={{ marginLeft: 30 }}>Loading...</h2>
 
@@ -42,7 +42,7 @@ const TrendsReChart = ({ chartSummaryData = [], chartData, asset }) => (
         <ComposedChart
           data={chartData}
           syncId='trends'
-          margin={{ top: 5, right: 5, left: 0, bottom: 5 }}
+          margin={{ top: 5, right: 36, left: 0, bottom: 5 }}
         >
           <XAxis
             dataKey='datetime'
@@ -75,7 +75,7 @@ const TrendsReChart = ({ chartSummaryData = [], chartData, asset }) => (
             yAxisId='axis-price'
             name={asset + '/USD'}
             dot={false}
-            strokeWidth={2}
+            strokeWidth={1.5}
             dataKey='priceUsd'
             stroke={ASSET_PRICE_COLOR}
           />
@@ -83,7 +83,7 @@ const TrendsReChart = ({ chartSummaryData = [], chartData, asset }) => (
             type='linear'
             dataKey={entity.index}
             dot={false}
-            strokeWidth={3}
+            strokeWidth={entity.index === 'merged' ? 1.5 : 3}
             name={entity.name}
             stroke={entity.color}
           />