+++ lib/sanbase/elasticsearch/elasticsearch_api_mock.ex
@@ -2,7 +2,7 @@ defmodule Sanbase.ElasticsearchMock do
   @behaviour Elasticsearch.API
 
   @impl true
-  def request(_config, :get, "/index1,index2,index3/_stats", _data, _opts) do
+  def request(_config, :get, "/index1,index2,index3,index4/_stats", _data, _opts) do
     {:ok,
      %HTTPoison.Response{
        status_code: 200,
@@ -11,7 +11,7 @@ defmodule Sanbase.ElasticsearchMock do
   end
 
   @impl true
-  def request(_config, :post, "/index1,index2,index3/_search", _data, _opts) do
+  def request(_config, :post, "/index1,index2,index3,index4/_search", _data, _opts) do
     {:ok,
      %HTTPoison.Response{
        status_code: 200,
@@ -38,6 +38,16 @@ defmodule Sanbase.ElasticsearchMock do
        body: %{
          "aggregations" => %{"subreddits" => %{"buckets" => [2, 4, 5, 6, 7, 8, 9, 0, 12, 2]}}
        }
+     }}
+  end
+
+  @impl true
+  def request(_config, :post, "/discord/_search", _data, _opts) do
+    # Buckets must be an Enumerable
+    {:ok,
+     %HTTPoison.Response{
+       status_code: 200,
+       body: %{"aggregations" => %{"channel_names" => %{"buckets" => [1, 2, 3, 4, 5, 6]}}}
      }}
   end
 end