+++ lib/sanbase/influxdb/measurement.ex
@@ -68,6 +68,18 @@ defmodule Sanbase.Influxdb.Measurement do
     end
   end
 
+  @doc ~s"""
+    convert a list of slugs to measurement-slug map
+  """
+  def names_from_slugs(slugs) when is_list(slugs) do
+    measurement_slug_map =
+      Project.tickers_by_slug_list(slugs)
+      |> Enum.map(fn {ticker, slug} -> {ticker <> "_" <> slug, slug} end)
+      |> Map.new()
+
+    {:ok, measurement_slug_map}
+  end
+
   # Private functions
 
   defp format_timestamp(%DateTime{} = datetime) do