+++ lib/sanbase_web/admin/model/exchange_address.ex
@@ -1,18 +1,18 @@
-defmodule Sanbase.ExAdmin.Model.ExchangeEthAddress do
+defmodule Sanbase.ExAdmin.Model.ExchangeAddress do
   use ExAdmin.Register
 
   alias Sanbase.Repo
-  alias Sanbase.Model.ExchangeEthAddress
+  alias Sanbase.Model.ExchangeAddress
 
-  register_resource ExchangeEthAddress do
-    form exchange_eth_address do
+  register_resource ExchangeAddress do
+    form exchange_address do
       inputs do
         content do
           """
           <div>
           Paste CSV in the following format:
           <ul>
-            <li>Format: <b>address*</b>,<b>name*</b>,<b>source</b>,<b>comments</b></li>
+            <li>Format: <b>address*</b>,<b>name*</b>,<b>source</b>,<b>comments</b>,<b>is dex</b>,<b>infrastructure id</b></li>
             <li>Required: address, name</li>
             <li>Column titles should be ommited</li>
             <li>Example:
@@ -23,7 +23,7 @@ defmodule Sanbase.ExAdmin.Model.ExchangeEthAddress do
           """
         end
 
-        input(exchange_eth_address, :csv, type: :text, label: "paste CSV")
+        input(exchange_address, :csv, type: :text, label: "paste CSV")
       end
     end
 
@@ -32,7 +32,7 @@ defmodule Sanbase.ExAdmin.Model.ExchangeEthAddress do
     end
   end
 
-  def process_csv(conn, %{exchange_eth_address: %{csv: csv}} = params) when not is_nil(csv) do
+  def process_csv(conn, %{exchange_address: %{csv: csv}} = params) when not is_nil(csv) do
     csv
     |> String.replace("\r", "")
     |> CSVLixir.Reader.read()
@@ -40,29 +40,33 @@ defmodule Sanbase.ExAdmin.Model.ExchangeEthAddress do
     |> Enum.map(&update_or_create_eth_address/1)
     |> Enum.each(&Repo.insert_or_update!/1)
 
-    {conn, %{exchange_eth_address: %{}}}
+    {conn, %{exchange_address: %{}}}
   end
 
   def process_csv(conn, params), do: {conn, params}
 
-  def update_or_create_eth_address([address, name, source, comments]) do
-    Repo.get_by(ExchangeEthAddress, address: address)
+  def update_or_create_eth_address([address, name, source, comments, is_dex, infrastructure_id]) do
+    Repo.get_by(ExchangeAddress, address: address)
     |> case do
       nil ->
-        %ExchangeEthAddress{}
-        |> ExchangeEthAddress.changeset(%{
+        %ExchangeAddress{}
+        |> ExchangeAddress.changeset(%{
           address: address,
           name: name,
           source: source,
-          comments: comments
+          comments: comments,
+          is_dex: is_dex,
+          infrastructure_id: infrastructure_id
         })
 
       exch_address ->
         exch_address
-        |> ExchangeEthAddress.changeset(%{
+        |> ExchangeAddress.changeset(%{
           name: name || exch_address.name,
           source: source || exch_address.source,
-          comments: comments || exch_address.comments
+          comments: comments || exch_address.comments,
+          is_dex: is_dex || exch_address.is_dex,
+          infrastructure_id: infrastructure_id || exch_address.infrastructure_id
         })
     end
   end