+++ lib/sanbase_web/graphql/schema.ex
@@ -71,6 +71,10 @@ defmodule SanbaseWeb.Graphql.Schema do
     ]
   end
 
+  def middleware(middlewares, field, object) do
+    SanbaseWeb.Graphql.Prometheus.Instrumenter.instrument(middlewares, field, object)
+  end
+
   query do
     @desc "Returns the user currently logged in."
     field :current_user, :user do
@@ -168,6 +172,30 @@ defmodule SanbaseWeb.Graphql.Schema do
 
       complexity(&Complexity.from_to_interval/3)
       cache_resolve(&PriceResolver.ohlc/3)
+    end
+
+    @desc ~s"""
+    Fetch data for each of the projects in the slugs lists
+    """
+    field :projects_list_stats, list_of(:project_stats) do
+      arg(:slugs, non_null(list_of(:string)))
+      arg(:from, non_null(:datetime))
+      arg(:to, non_null(:datetime))
+
+      cache_resolve(&PriceResolver.multiple_projects_stats/3)
+    end
+
+    @desc ~s"""
+    Fetch data bucketed by interval. The returned marketcap and volume are the sum
+    of the marketcaps and volumes of all projects for that given time interval
+    """
+    field :projects_list_history_stats, list_of(:combined_projects_stats) do
+      arg(:slugs, non_null(list_of(:string)))
+      arg(:from, non_null(:datetime))
+      arg(:to, non_null(:datetime))
+      arg(:interval, non_null(:string))
+
+      cache_resolve(&ProjectResolver.combined_history_stats/3)
     end
 
     @desc "Returns a list of available github repositories."
@@ -476,11 +504,11 @@ defmodule SanbaseWeb.Graphql.Schema do
       * interval - an integer followed by one of: `m`, `h`, `d`, `w`
       * from - a string representation of datetime value according to the iso8601 standard, e.g. "2018-04-16T10:02:19Z"
       * to - a string representation of datetime value according to the iso8601 standard, e.g. "2018-04-16T10:02:19Z"
-      * socialVolumeType - one of the following:
-        1. PROFESSIONAL_TRADERS_CHAT_OVERVIEW
-        2. TELEGRAM_CHATS_OVERVIEW
-        3. TELEGRAM_DISCUSSION_OVERVIEW
-        It is used to select the source of the mentions count.
+      * socialVolumeType - the source of mention counts, one of the following:
+        1. "PROFESSIONAL_TRADERS_CHAT_OVERVIEW" - shows how many times the given project has been mentioned in the professional traders chat
+        2. "TELEGRAM_CHATS_OVERVIEW" - shows how many times the given project has been mentioned across all telegram chats, except the project's own community chat (if there is one)
+        3. "TELEGRAM_DISCUSSION_OVERVIEW" - the general volume of messages in the project's community chat (if there is one)
+        4. "DISCORD_DISCUSSION_OVERVIEW" - shows how many times the given project has been mentioned in the discord channels
     """
     field :social_volume, list_of(:social_volume) do
       arg(:slug, non_null(:string))
@@ -510,6 +538,7 @@ defmodule SanbaseWeb.Graphql.Schema do
         1. TELEGRAM
         2. PROFESSIONAL_TRADERS_CHAT
         3. REDDIT
+        4. DISCORD
       * searchText - a string containing the key words for which the sources should be searched.
       * interval - an integer followed by one of: `m`, `h`, `d`, `w`
       * from - a string representation of datetime value according to the iso8601 standard, e.g. "2018-04-16T10:02:19Z"
@@ -568,6 +597,17 @@ defmodule SanbaseWeb.Graphql.Schema do
     @desc "Fetch all public favourites lists"
     field :fetch_all_public_user_lists, list_of(:user_list) do
       resolve(&UserListResolver.fetch_all_public_user_lists/3)
+    end
+
+    @desc ~s"""
+    Fetch public favourites list by list id.
+    If the list is owned by the current user then the list can be private as well.
+    This query returns either a single user list item or null.
+    """
+    field :user_list, :user_list do
+      arg(:user_list_id, non_null(:id))
+
+      resolve(&UserListResolver.user_list/3)
     end
 
     @desc "Returns statistics for the data stored in elasticsearch"