+++ lib/sanbase_web/graphql/schema/tech_indicators_types.ex
@@ -57,12 +57,14 @@ defmodule SanbaseWeb.Graphql.TechIndicatorsTypes do
     value(:professional_traders_chat_overview)
     value(:telegram_chats_overview)
     value(:telegram_discussion_overview)
+    value(:discord_discussion_overview)
   end
 
   enum :topic_search_sources do
     value(:telegram)
     value(:professional_traders_chat)
     value(:reddit)
+    value(:discord)
   end
 
   object :topic_search do