+++ app/CHANGELOG.md
@@ -1,5 +1,16 @@
 ## CHANGELOG
 
+- 2018-22-10 | v.0.19.3 'Phobos':
+
+  Upgrade Metamask
+  Update Assets Menu (Added Watchlists)
+  Show Watchlists for anonymous users
+  Small style updates
+
+- 2018-12-10 | v.0.19.2 'Phobos':
+
+  Feedback form send feedbacks to our private Discord channel
+
 - 2018-12-10 | v.0.19.1 'Phobos':
 
   Fix Trends Stats